# packages/api-client/openapi.json
{"openapi": "3.1.0", "info": {"title": "Glia API", "description": "Realtime transcription and visual-intent API for Glia.", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}, "/api/realtime-token": {"post": {"summary": "Realtime Token", "operationId": "realtime_token_api_realtime_token_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RealtimeTokenRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RealtimeTokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/ledger": {"get": {"summary": "Ledger", "description": "Credits spent this process. In-process, so it resets with the API \u2014 that is a known\nlimitation of today's counter, not a claim about the monthly pool.", "operationId": "ledger_v1_ledger_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LedgerSnapshot"}}}}}}}, "/v1/discover": {"post": {"summary": "Discover", "description": "Resolve the spoken subject against Cala, then find the sources that answer it.\n\nThree gates stand in front of the two upstream calls, and a settled turn that fails any of\nthem costs nothing: the per-session debounce replays the last answer, the cache replays a\nprevious one, and the ledger refuses outright once the budget is gone.", "operationId": "discover_v1_discover_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiscoverRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiscoverResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/image": {"get": {"summary": "Image Proxy", "description": "Re-serve a discovered image so the browser never hot-links a third party.\n\n`url` is remote-supplied and is treated as such: ImageFetcher owns the\nscheme and host allowlists, the resolved-address check, redirect refusal,\nthe content-type check and the streamed byte cap.", "operationId": "image_proxy_api_image_get", "parameters": [{"name": "url", "in": "query", "required": true, "schema": {"type": "string", "maxLength": 2000, "title": "Url"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/generate": {"post": {"summary": "Generate", "description": "Transcript + intent + pins \u2192 one prompt \u2192 one image.\n\nTwo properties are load-bearing and everything else here serves them.\n\nThe pins are a real input. Every pin's title reaches the synthesis, and every pin carrying\nan origin image URL is additionally fetched here and re-hosted on fal's storage, so that\nwhat goes out as `image_urls` is always a URL fal can read. Our own `/api/image` proxy is\nnever sent: fal cannot reach localhost, and the pin's display URL is not its origin.\n\nA pin that cannot be fetched or uploaded is dropped by id rather than fatal. The generation\nruns on the pins that worked, `reference_count` reports what actually conditioned the image,\nand `unavailable_references` names the ones that did not.\n\nThe prompt is a deliverable. It is returned verbatim on every outcome that reached fal,\nincluding the timeout, because it is the answer to \"here is what we understood you to\nmean\" and that answer survives a generation that did not land.", "operationId": "generate_v1_generate_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CalaEntityHit": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "entity_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Entity Type"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["id", "name"], "title": "CalaEntityHit", "description": "One row of `GET /entities?name=\u2026` \u2014 the entity-resolution step."}, "CalaExplanation": {"properties": {"content": {"type": "string", "title": "Content", "default": ""}, "references": {"items": {"type": "string"}, "type": "array", "title": "References"}}, "type": "object", "title": "CalaExplanation"}, "CalaNamedUrl": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Url"}}, "type": "object", "title": "CalaNamedUrl"}, "CalaOrigin": {"properties": {"source": {"anyOf": [{"$ref": "#/components/schemas/CalaNamedUrl"}, {"type": "null"}]}, "document": {"anyOf": [{"$ref": "#/components/schemas/CalaNamedUrl"}, {"type": "null"}]}}, "type": "object", "title": "CalaOrigin"}, "CalaSearchEntity": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "entity_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Entity Type"}, "mentions": {"items": {"type": "string"}, "type": "array", "title": "Mentions"}}, "type": "object", "required": ["id", "name"], "title": "CalaSearchEntity"}, "DiscoverRequest": {"properties": {"transcript": {"type": "string", "maxLength": 50000, "minLength": 1, "title": "Transcript"}, "session_id": {"type": "string", "maxLength": 128, "minLength": 1, "title": "Session Id"}}, "additionalProperties": false, "type": "object", "required": ["transcript", "session_id"], "title": "DiscoverRequest"}, "DiscoverResponse": {"properties": {"status": {"type": "string", "enum": ["ok", "empty", "rate_limited", "budget_exhausted"], "title": "Status"}, "session_id": {"type": "string", "title": "Session Id"}, "subject": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subject"}, "query": {"type": "string", "title": "Query"}, "entity": {"anyOf": [{"$ref": "#/components/schemas/CalaEntityHit"}, {"type": "null"}]}, "answer": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Answer"}, "explainability": {"items": {"$ref": "#/components/schemas/CalaExplanation"}, "type": "array", "title": "Explainability"}, "context": {"items": {"$ref": "#/components/schemas/EvidenceItem"}, "type": "array", "title": "Context"}, "entities": {"items": {"$ref": "#/components/schemas/CalaSearchEntity"}, "type": "array", "title": "Entities"}, "cached": {"type": "boolean", "title": "Cached"}, "ledger": {"$ref": "#/components/schemas/LedgerSnapshot"}, "correlation_id": {"type": "string", "title": "Correlation Id"}}, "additionalProperties": false, "type": "object", "required": ["status", "session_id", "subject", "query", "entity", "answer", "explainability", "context", "entities", "cached", "ledger", "correlation_id"], "title": "DiscoverResponse"}, "EvidenceItem": {"properties": {"id": {"type": "string", "title": "Id"}, "content": {"type": "string", "title": "Content"}, "origins": {"items": {"$ref": "#/components/schemas/CalaOrigin"}, "type": "array", "title": "Origins"}, "carried_answer": {"type": "boolean", "title": "Carried Answer"}}, "additionalProperties": false, "type": "object", "required": ["id", "content", "origins", "carried_answer"], "title": "EvidenceItem", "description": "A `context[]` item, with the explainability join already resolved.\n\n`carried_answer` is ours, not Cala's: it is true when this item's id appears in some\n`explainability[].references`, which is the salience signal the UI badges."}, "GenerateRequest": {"properties": {"session_id": {"type": "string", "maxLength": 128, "minLength": 1, "title": "Session Id"}, "transcript": {"type": "string", "maxLength": 50000, "minLength": 1, "title": "Transcript"}, "pins": {"items": {"$ref": "#/components/schemas/PinnedRef"}, "type": "array", "maxItems": 32, "title": "Pins"}}, "additionalProperties": false, "type": "object", "required": ["session_id", "transcript"], "title": "GenerateRequest"}, "GenerateResponse": {"properties": {"status": {"type": "string", "enum": ["ok", "timeout", "already_generating", "reference_unavailable"], "title": "Status"}, "session_id": {"type": "string", "title": "Session Id"}, "image_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Image Url"}, "prompt": {"type": "string", "title": "Prompt"}, "model": {"type": "string", "title": "Model"}, "reference_count": {"type": "integer", "title": "Reference Count"}, "unavailable_references": {"items": {"type": "string"}, "type": "array", "title": "Unavailable References"}, "correlation_id": {"type": "string", "title": "Correlation Id"}}, "additionalProperties": false, "type": "object", "required": ["status", "session_id", "image_url", "prompt", "model", "reference_count", "correlation_id"], "title": "GenerateResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "service": {"type": "string", "title": "Service"}, "mode": {"type": "string", "title": "Mode"}, "realtime": {"type": "string", "title": "Realtime"}, "distiller": {"type": "string", "title": "Distiller"}, "image_discovery": {"type": "boolean", "title": "Image Discovery", "description": "Whether this server can return image candidates for a distilled intent. False means the client must not offer an image-bearing output mode."}}, "additionalProperties": false, "type": "object", "required": ["status", "service", "mode", "realtime", "distiller", "image_discovery"], "title": "HealthResponse"}, "LedgerSnapshot": {"properties": {"budget": {"type": "integer", "title": "Budget"}, "spent": {"type": "integer", "title": "Spent"}, "remaining": {"type": "integer", "title": "Remaining"}, "search_calls": {"type": "integer", "title": "Search Calls"}, "entity_calls": {"type": "integer", "title": "Entity Calls"}}, "additionalProperties": false, "type": "object", "required": ["budget", "spent", "remaining", "search_calls", "entity_calls"], "title": "LedgerSnapshot", "description": "Credits spent this process. `entity_calls` is counted separately because\ndocs/PARTNERS.md flags it as undocumented whether `/entities` consumes credits; we charge\nit against the budget (the safe direction) while keeping it measurable against the console."}, "PinnedRef": {"properties": {"id": {"type": "string", "maxLength": 128, "minLength": 1, "title": "Id"}, "title": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Title"}, "lane": {"type": "string", "maxLength": 64, "minLength": 1, "title": "Lane"}, "image_url": {"anyOf": [{"type": "string", "maxLength": 2048}, {"type": "null"}], "title": "Image Url"}, "origin_image_url": {"anyOf": [{"type": "string", "maxLength": 2048}, {"type": "null"}], "title": "Origin Image Url"}, "source_url": {"anyOf": [{"type": "string", "maxLength": 2048}, {"type": "null"}], "title": "Source Url"}}, "additionalProperties": false, "type": "object", "required": ["id", "title", "lane"], "title": "PinnedRef"}, "RealtimeTokenRequest": {"properties": {"client_id": {"type": "string", "maxLength": 128, "minLength": 8, "title": "Client Id"}, "languages": {"items": {"type": "string", "enum": ["ca", "en", "es", "fr"]}, "type": "array", "maxItems": 4, "minItems": 1, "title": "Languages"}}, "additionalProperties": false, "type": "object", "required": ["client_id"], "title": "RealtimeTokenRequest"}, "RealtimeTokenResponse": {"properties": {"value": {"type": "string", "title": "Value"}, "expires_at": {"type": "integer", "title": "Expires At"}, "model": {"type": "string", "title": "Model"}}, "additionalProperties": false, "type": "object", "required": ["value", "expires_at", "model"], "title": "RealtimeTokenResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}